financial-rag-kg/requirements.txt
# Phase 1: Data Collection
requests>=2.31.0
newspaper3k>=0.2.8  # For advanced article parsing (optional)
python-dotenv>=1.0.0  # For environment variables

# Phase 2: Knowledge Graph (upcoming)
neo4j>=5.14.0
llama-index>=0.9.0
openai>=1.0.0
langchain>=0.1.0

# Phase 3: RAG Pipeline (upcoming)
chromadb>=0.4.0  # Vector database
sentence-transformers>=2.2.0
tiktoken>=0.5.0

# Phase 4: Streamlit Interface (upcoming)
streamlit>=1.28.0
streamlit-agraph>=0.0.45  # Graph visualization
plotly>=5.17.0
pandas>=2.1.0

# Utilities
python-dateutil>=2.8.0
tqdm>=4.66.0

# Testing & Development
pytest>=7.4.0
black>=23.10.0
flake8>=6.1.0
